--- src/background/js/defaultSettings.json
@@ -1,12 +1,0 @@
-{
-    "token": null,
-    "autoUpdate": true,
-    "autoRefresh": 60,
-    "listItemOfType": "all",
-    "numberOfItems": 4,
-    "updateFavicon": true,
-    "viewportSide": "left",
-    "sortBy": "CREATED_AT",
-    "repos": [],
-    "theme": "dark"
-}
