+++ src/background/lib/ports.js
@@ -1,6 +1,7 @@
 let index = 0;
 const ports = {};
 
+// TODO: Change to an object instead of multiple exports?
 export function add(port) {
 	const id = index;
 	ports[id] = port;
@@ -18,4 +19,9 @@ export function messageAll(message) {
 	for (const id in ports) {
 		ports[id].postMessage(message);
 	}
+}
+
+// TOdo: Trengs denne egentlig? Eller baree bruke messageAll?
+export function sendToAllTabs(data) {
+	messageAll(data);
 }
