+++ src/background/settings/__tests__/quickStorage.test.js
