+++ src/background/settings/defaultSettings.json
@@ -1,0 +1,12 @@
+{
+	"token": null,
+	"autoUpdate": true,
+	"autoRefresh": 60,
+	"listItemOfType": "all",
+	"numberOfItems": 4,
+	"updateFavicon": true,
+	"viewportSide": "left",
+	"sortBy": "CREATED_AT",
+	"repos": [],
+	"theme": "dark"
+}
