+++ src/background/settings/quickStorage.js
